requirements.txt.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

easyocr==1.7.0
opencv-python-headless==4.8.1.78
pillow==10.1.0

deep-translator==1.11.4

pandas==2.1.3
openpyxl==3.1.2

scikit-learn==1.3.2
imbalanced-learn==0.11.0

torch
torchvision
torchaudio

transformers==4.30.2
sentence-transformers==2.2.2
huggingface-hub==0.14.1
